requirements.txt
aiohttp==3.7.4
aiosqlite==0.17.0
argon2-cffi==21.1.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
bitstring==3.1.9
blspy==1.0.6
cffi==1.15.0
chardet==3.0.4
chia-blockchain==1.2.6
chia-dev-tools==1.0.4
chiabip158==1.0
chiapos==1.0.4
chiavdf==1.0.3
click==7.1.2
clvm==0.9.7
clvm-rs==0.1.11
clvm-tools==0.4.3
colorama==0.4.4
colorlog==5.0.1
concurrent-log-handler==0.9.19
cryptography==3.4.7
dnspython==2.1.0
fasteners==0.16.3
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
keyring==23.0.1
keyrings.cryptfile==1.3.4
multidict==5.2.0
packaging==21.2
pluggy==1.0.0
portalocker==2.3.2
py==1.11.0
pybind11==2.8.1
pycparser==2.21
pycryptodome==3.11.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
pytimeparse==1.1.8
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==5.4.1
setproctitle==1.2.2
six==1.16.0
sortedcontainers==2.3.0
toml==0.10.2
typing-extensions==3.10.0.2
watchdog==2.1.3
websockets==8.1
yarl==1.7.2
zipp==3.6.0
